python_backend/requirements.txt
django>=5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.0
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0
